analytics/terraform/spark-k8s-operator/examples/spark-sleep/locust/requirements.txt
blinker==1.8.2
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
google-auth==2.33.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==30.1.0
locust==2.31.2
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==72.1.0
six==1.16.0
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.0.1